refactor(agent): split agent-model.ts and AgentLaunchModal.tsx
Tier 2 item from docs/reports/REPORT_LARGE_FILE_MODULARIZATION_SCAN_2026_07_22.md.

agent-model.ts: AgentViewModel class (viewIcon/viewName/launch orchestration,
one cohesive ViewModel) vs. pure, zero-coupling config-file synthesis helpers
(buildConfigFiles/expandTemplate/buildSettingsWithHooks/buildMcpConfig) and
launch-env helpers (checkNodejsForProvider/agentmuxHome/resolveCliDir).
Extracted to agent-config-builder.ts + agent-launch-env.ts. 974 -> 680 lines.

AgentLaunchModal.tsx: extracted "Continue / New" mode logic to
hooks/useContinueOrNewMode.ts and auth-gating logic to hooks/
useLaunchAuthGate.ts. flow (the createLaunchFlowStore store) is threaded by
reference into both hooks, never destructured, to preserve SolidJS
reactivity. 987 -> 844 lines.

tsc --noEmit clean; AgentLaunchModal.integration.test.tsx and the full
frontend/app/view/agent test sweep (720/721, one pre-existing unrelated
flaky timeout confirmed passing in isolation) pass unchanged.

<!-- agentmux:agent_id=agentx -->
